FAQ integration and adjustments
